Created a 3D rendering engine that reads obj files
Engine is capable of rendering using both an orthographic camera and a
perspective camera. Reads basic .obj format including an encoding for
how to scale the mesh (# s [scale value]). When reading .obj files, it's
easiest to just start up a SimpleHTTPServer with python to view them.
You can switch between cameras with 'c', rotate the camera with 'a' and
'd', and move the camera forward and backward with 'w' and s',
respectively.